Fix bug in catch image uploads
-Create directory and index.js file for CreateCatchForm
-Move all image handling from UserCatchesMap to CreateCatchForm
-Fix CreateCatchForm/UserCatchesMap so it retains values when location is re-selected on activated form
-Move generateFileDataArray to utilities/helpers
